raster-checker command removed, Schematisation Algorithm refactored
