+++ constraints.txt
@@ -1,0 +1,22 @@
+# Generated by dependencies.py
+SQLAlchemy<1.4
+GeoAlchemy2==0.10.2
+lizard-connector==0.7.3
+pyqtgraph>=0.11.1,<0.12
+threedigrid>=1.2.3
+cached-property
+threedi-modelchecker==0.31
+threedidepth==0.4
+click>=8.0
+alembic==1.8.*
+importlib-resources
+zipp
+Mako
+netCDF4
+cftime
+packaging
+colorama
+networkx
+condenser>=0.1.1
+pygeos>=0.12.0
+threedigrid-builder>=1.3.6
