+++ processing/deps/raster_checker/raster_checker_log.py
@@ -1,7 +1,7 @@
 # (c) Nelen & Schuurmans, see LICENSE.rst.
 from cached_property import cached_property
 from jinja2 import Template
-from ThreeDiToolbox.tool_commands.raster_checker.constants import RASTER_CHECKER_MAPPER
+from ThreeDiToolbox.processing.deps.raster_checker.constants import RASTER_CHECKER_MAPPER
 from ThreeDiToolbox.utils.user_messages import StatusProgressBar
 
 import logging
