+++ processing/deps/raster_checker/raster_checker_prework.py
@@ -2,10 +2,10 @@
 from cached_property import cached_property
 from sqlalchemy import select
 from sqlalchemy import Table
-from ThreeDiToolbox.tool_commands.raster_checker.constants import (
+from ThreeDiToolbox.processing.deps.raster_checker.constants import (
     NON_SETTINGS_TBL_WITH_RASTERS,
 )
-from ThreeDiToolbox.tool_commands.raster_checker.constants import V2_TABLES
+from ThreeDiToolbox.processing.deps.raster_checker.constants import V2_TABLES
 
 import logging
